refactor(usage): fold native-commands reference into Analytics, drop standalone page
The standalone /usage page mostly re-presented data Analytics already
shows — its 'Delegation & ecosystem' section (subagents by cost, skills,
MCP servers) and per-model breakdown are richer than what the page had.
The only piece unique to it was the per-agent reference: which in-session
command each agent uses for usage, and its limit/reset model.

So: remove the page, its nav link, and the now-unused GET /usage backend
endpoint (+ contribution/by-model helpers, and the history_store
first_ts/last_ts fields that only served them). Add a compact
'Native commands & limits' section at the bottom of Analytics that keeps
the curated reference — where each agent's live remaining quota actually
lives, since TT only re-counts local tokens.

Qwen entry uses @Pauliehedron's tested v0.19.6 findings (native /context,
/statusline, /summary, /recap, /export; sessions at ~/.qwen/…/chats/*.jsonl),
which also match QwenLM/qwen-code source. Vibe corrected: it has no
/status (only --max-tokens/--max-price flags).
