+++ backend/main.py
@@ -6286,219 +6286,6 @@ async def get_analytics(
     }
 
 
-_USAGE_WINDOWS = ("today", "7d", "30d", "all")
-
-# Threshold for the "long-running session" contribution line. Claude Code's own
-# /usage flags 8h+ sessions (background/loop runs) as a usage driver; we mirror
-# that so the cross-agent panel reads the same way.
-_LONG_SESSION_HOURS = 8.0
-def _usage_duration_h(s: Dict[str, Any]) -> float:
-    """Session wall-clock hours from first_ts..last_ts when both are known.
-    Returns 0.0 for sessions missing a start bound (live-only rows not yet
-    persisted), so they never inflate the long-session numerator."""
-    a, b = s.get("first_ts"), s.get("last_ts")
-    if not a or not b:
-        return 0.0
-    try:
-        return max(0.0, (datetime.fromisoformat(b) - datetime.fromisoformat(a)).total_seconds() / 3600.0)
-    except (ValueError, TypeError):
-        return 0.0
-
-
-def _usage_is_subagent_heavy(s: Dict[str, Any]) -> bool:
-    """True when the session spawned subagents (Task/Agent delegation). Each
-    subagent runs its own requests, so these sessions drive usage disproportionately."""
-    d = s.get("delegation")
-    if isinstance(d, dict):
-        if (d.get("spawn_count") or 0) > 0:
-            return True
-        by_type = d.get("by_type")
-        if isinstance(by_type, dict) and any((v or 0) > 0 for v in by_type.values()):
-            return True
-    si = s.get("subagent_info")
-    if isinstance(si, dict) and (si.get("spawn_count") or si.get("count") or 0) > 0:
-        return True
-    if isinstance(si, list) and si:
-        return True
-    return False
-
-
-def _usage_contributions(sessions: List[Dict[str, Any]]) -> Dict[str, Any]:
-    """The cross-agent analogue of Claude Code /usage's "What's contributing to
-    your limits usage?" panel, computed from local session records: behavioral
-    drivers (subagent-heavy / long-running) plus per-entity breakdown tables
-    (agents, skills, subagents, MCP servers). Shares are cost-weighted."""
-    total_cost = sum(float(s.get("cost", 0.0) or 0.0) for s in sessions)
-    subagent_cost = sum(float(s.get("cost", 0.0) or 0.0) for s in sessions if _usage_is_subagent_heavy(s))
-    long_cost = sum(float(s.get("cost", 0.0) or 0.0) for s in sessions if _usage_duration_h(s) >= _LONG_SESSION_HOURS)
-
-    by_agent: Dict[str, float] = {}
-    skills: Dict[str, int] = {}
-    subagents: Dict[str, int] = {}
-    mcp: Dict[str, int] = {}
-    for s in sessions:
-        cost = float(s.get("cost", 0.0) or 0.0)
-        by_agent[s.get("agent", "unknown")] = by_agent.get(s.get("agent", "unknown"), 0.0) + cost
-        for sk in (s.get("skills_used") or []):
-            if isinstance(sk, dict) and sk.get("name"):
-                skills[sk["name"]] = skills.get(sk["name"], 0) + int(sk.get("count", 1) or 1)
-        d = s.get("delegation")
-        if isinstance(d, dict) and isinstance(d.get("by_type"), dict):
-            for t, n in d["by_type"].items():
-                # by_type values are either a bare count or a {count/spawn_count/...} dict.
-                if isinstance(n, dict):
-                    cnt = n.get("count") or n.get("spawn_count") or n.get("calls") or 0
-                else:
-                    cnt = n
-                try:
-                    subagents[t] = subagents.get(t, 0) + int(cnt or 0)
-                except (TypeError, ValueError):
-                    continue
-        mu = s.get("mcp_usage")
-        if isinstance(mu, dict):
-            for server, tools in mu.items():
-                n = sum(tools.values()) if isinstance(tools, dict) else int(tools or 0)
-                mcp[server] = mcp.get(server, 0) + n
-
-    def _pct(part: float) -> Optional[float]:
-        return round(100.0 * part / total_cost, 1) if total_cost > 0 else None
-
-    def _top(d: Dict[str, Any], n: int = 5) -> List[Dict[str, Any]]:
-        return [{"name": k, "value": v} for k, v in sorted(d.items(), key=lambda kv: (-kv[1], kv[0]))[:n]]
-
-    return {
-        "session_count": len(sessions),
-        "total_cost": round(total_cost, 6),
-        "subagent_share": _pct(subagent_cost),
-        "long_session_share": _pct(long_cost),
-        # Two Claude /usage drivers we deliberately don't fake: ">150k context"
-        # needs a per-session context-window peak we don't record, and "4+
-        # sessions in parallel" needs per-request timing (whole-session interval
-        # overlap saturates to ~100% for long loop sessions). Left null; the UI
-        # notes they aren't reconstructed locally.
-        "context_share": None,
-        "parallel_share": None,
-        "by_agent": [{"name": k, "value": round(v, 6)} for k, v in sorted(by_agent.items(), key=lambda kv: (-kv[1], kv[0]))],
-        "skills": _top(skills),
-        "subagents": _top(subagents),
-        "mcp": _top(mcp),
-    }
-
-
-@app.get("/usage")
-async def get_usage():
-    """Unified equivalent of each agent's own in-session `/usage` command:
-    per-agent tokens/cost/cache-hit/session-count for today (local calendar
-    day), rolling 7d, rolling 30d, and all-time, plus the same plan/quota
-    context Settings' drain-order view already computes. One place to see
-    what every agent's native /usage, /stats, or /status would show, instead
-    of switching tools."""
-    import history_store
-    from billing_mode import get_all as get_billing_modes
-    from billing_route import get_route_overview, load_plans, DEFAULT_PLAN
-
-    now_local = datetime.now().astimezone()
-    today_start = now_local.replace(hour=0, minute=0, second=0, microsecond=0)
-    window_bounds = {
-        "today": today_start.astimezone(timezone.utc).isoformat(),
-        "7d": (now_local - timedelta(days=7)).astimezone(timezone.utc).isoformat(),
-        "30d": (now_local - timedelta(days=30)).astimezone(timezone.utc).isoformat(),
-        "all": None,
-    }
-
-    stored = history_store.query(None, None, [], [], [])
-    merged: Dict[tuple, Dict[str, Any]] = {(s["agent"], s["id"]): s for s in stored}
-    for s in await get_sessions_cached():
-        key = (s.get("agent"), s.get("id"))
-        prev = merged.get(key)
-        # The live scan wins for token/cost figures, but doesn't carry the
-        # durable window bounds (first_ts/last_ts) history persists. Backfill
-        # them so session-duration / concurrency drivers below aren't zeroed
-        # out for sessions that also exist in history. Copy so we never mutate
-        # a cached live-session object.
-        if prev is not None and (not s.get("first_ts") or not s.get("last_ts")):
-            s = {**s}
-            for k in ("first_ts", "last_ts"):
-                if not s.get(k) and prev.get(k):
-                    s[k] = prev[k]
-        merged[key] = s
-    sessions = list(merged.values())
-
-    def _empty_window() -> Dict[str, Any]:
-        return {"input": 0, "output": 0, "cached": 0, "total": 0, "cost": 0.0,
-                "session_count": 0, "_models": {}}
-
-    # Sessions that fall inside each window, kept so the cross-agent
-    # "what's driving usage" panel can be computed per window (below).
-    windowed: Dict[str, List[Dict[str, Any]]] = {w: [] for w in _USAGE_WINDOWS}
-
-    by_agent: Dict[str, Dict[str, Dict[str, Any]]] = {}
-    for s in sessions:
-        agent = s.get("agent")
-        if not agent:
-            continue
-        ts = s.get("timestamp")
-        ts_iso = ts.astimezone(timezone.utc).isoformat() if isinstance(ts, datetime) else None
-        rows = by_agent.setdefault(agent, {w: _empty_window() for w in _USAGE_WINDOWS})
-        st = s.get("tokens", {})
-        scost = s.get("cost", 0.0) or 0.0
-        model = s.get("model") or "unknown"
-        for w, bound in window_bounds.items():
-            if bound is not None and (ts_iso is None or ts_iso < bound):
-                continue
-            row = rows[w]
-            for k in ("input", "output", "cached", "total"):
-                row[k] += st.get(k, 0)
-            row["cost"] += scost
-            row["session_count"] += 1
-            m = row["_models"].setdefault(model, {"input": 0, "output": 0, "cached": 0, "total": 0, "cost": 0.0, "session_count": 0})
-            for k in ("input", "output", "cached", "total"):
-                m[k] += st.get(k, 0)
-            m["cost"] += scost
-            m["session_count"] += 1
-            windowed[w].append(s)
-
-    for rows in by_agent.values():
-        for row in rows.values():
-            row["cache_hit_pct"] = _cache_hit_pct(row["input"], row["cached"])
-            row["cost"] = round(row["cost"], 6)
-            # Per-model breakdown (input/output/cache/cost), biggest first — the
-            # table every agent's own /cost, /usage or /stats shows.
-            models = row.pop("_models")
-            for m in models.values():
-                m["cache_hit_pct"] = _cache_hit_pct(m["input"], m["cached"])
-                m["cost"] = round(m["cost"], 6)
-            row["by_model"] = [
-                {"model": name, **vals}
-                for name, vals in sorted(models.items(), key=lambda kv: (-kv[1]["cost"], -kv[1]["total"], kv[0]))
-            ]
-
-    contributions = {w: _usage_contributions(windowed[w]) for w in _USAGE_WINDOWS}
-
-    agents = sorted(set(_list_available_agents()) | set(by_agent.keys()))
-    modes = get_billing_modes(agents)
-    plans = load_plans()
-    billing = {
-        a: get_route_overview(a, plan=plans.get(a, DEFAULT_PLAN), mode=modes.get(a, {}).get("mode"))
-        for a in agents
-    }
-
-    def _empty_agent() -> Dict[str, Dict[str, Any]]:
-        out = {}
-        for w in _USAGE_WINDOWS:
-            row = _empty_window(); row.pop("_models")
-            row["cache_hit_pct"] = None; row["by_model"] = []
-            out[w] = row
-        return out
-
-    return {
-        "agents": {a: by_agent.get(a, _empty_agent()) for a in agents},
-        "billing": billing,
-        "contributions": contributions,
-        "generated_at": datetime.now(timezone.utc).isoformat(),
-    }
-
-
 def _parse_skill_md(p: Path):
     """Read SKILL.md frontmatter; return {name, description}."""
     try:
