+++ frontend/src/components/Navigation.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { usePathname } from "next/navigation";
 import {
   LayoutDashboard, Folder, BarChart3, Activity, Settings2,
-  PanelLeftOpen, PanelLeftClose, Zap, Gauge,
+  PanelLeftOpen, PanelLeftClose, Zap,
 } from "lucide-react";
 import { useResource } from "@/lib/api";
 import { ALL_AGENT_KEYS, getAgent } from "@/lib/agents";
@@ -22,7 +22,6 @@ const LINKS = [
   { name: "Dashboard", href: "/",         icon: LayoutDashboard },
   { name: "Projects",  href: "/projects", icon: Folder },
   { name: "Analytics", href: "/analytics", icon: BarChart3 },
-  { name: "Usage", href: "/usage", icon: Gauge },
   { name: "Local Models", href: "/local-models", icon: Zap },
 ];
